requirements.txt
graphrag
pandas
numpy
networkx
obsidiantools
python-frontmatter
fastmcp
uvicorn
openai
python-dotenv
pyyaml
scikit-learn
watchdog 